requirements.txt
streamlit
openai-whisper
pandas
scikit-learn
torch
